+++ pandas_ml/__init__.py
@@ -3,3 +3,4 @@
 from pandas_ml.core import ModelFrame, ModelSeries       # noqa
 from pandas_ml.tools import info                         # noqa
 from pandas_ml.version import version as __version__     # noqa
+import pandas_ml.compat as compat                        # noqa