+++ pandas_ml/compat.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python
+
+from distutils.version import LooseVersion
+
+
+def _SKLEARN_ge_017():
+    import sklearn
+    return sklearn.__version__ >= LooseVersion('0.17.0')